+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
   - sfdx force:data:bulk:delete -> sf data delete bulk
   - sfdx alias:list -> sf alias list
   - sfdx alias:set -> sf alias set
+  - sfdx force:apex:test:run -> sf apex run test
+  - sfdx force:apex:execute -> sf apex run
   - sfdx hardis:* -> sf hardis:*
 - Get rid of sfdx-essentials dependency by internalizing its used commands
   - sf hardis:packagexml:append